Adjust Puppetfile functionality
Previously nubis-builder would use *either* the included puppet modules from nubis-base *or* it would use the puppet modules from your local repository.

This changes that behavior so that what is declared in your local Puppetfile will be *added* to (or overwrite) the /etc/puppet/modules on the instance as it is provisioned.

For example in this [Puppetfile](https://github.com/bhourigan/nubis-akamai-dns/blob/177062f6d333e7296f302180aeb11be0a557b38b/nubis/Puppetfile) the author is including a custom module but is also has to include the puppetlabs apache module. With this change the author would only need to include the custom module as the apache module is already shipped with nubis-base.

This change also removes the need to have two seperate puppet-masterless provisioners as nubis-base can 'puppet apply' from the instance instead of the local developer filesystem.